+++ src/controllers/messagesController.js
@@ -5,6 +5,7 @@ const multerConfig = require("../libs/multerConfig");
 const { unlink } = require("fs-extra");
 const { httpError } = require("../helpers/handleError");
 const path = require("path");
+const fs = require("fs");
 const validations = require("../validators/info");
 
 const parseId = (id) => {
@@ -126,7 +127,9 @@ exports.deleteMessagesById = async (req, res) => {
         .send({ message: "El mensaje que desea borrar no existe" });
 
     if (doc.imgURL != "ifgf.png") {
-      unlink(path.resolve("./public/uploads/" + doc.imgURL));
+      if (fs.existsSync(path.resolve("./public/uploads/" + doc.imgURL))) {
+        unlink(path.resolve("./public/uploads/" + doc.imgURL));
+      }
     }
     res.send({ message: "Eliminado con exito" });
   } catch (error) {
